Use non-interactive registration, which simplifies setup
It should be noted that the GITLAB_RUNNER_TOKEN now denotes the
registration token from Gitlab, instead of the runner token that gets
inserted into /etc/gitlab-runner/config.toml after register completes.
This has the effect of runners being recognized by Gitlab as fresh ones,
whereas previously a runner could survive a docker-compose down so long
as the same token was passed in.